chore: rename type.ts to types.ts
