+++ src/components/apps/Calendar/context.ts
@@ -1,4 +1,4 @@
 import { createContext } from 'preact';
-import { ICalendarAppContext } from './type';
+import { ICalendarAppContext } from './types';
 
 export const CalendarAppContext = createContext<ICalendarAppContext>({} as ICalendarAppContext);
